add the x509 certificate from the Crypki server also to the Key info
This is required, otherwise Keycloak can't build up the JWKS document which allows clients to validate the token signature signed with this key.

To get this working in a performant way, I also refactored the service and provider classes a bit (only create the service once for the lifetime of the class, not explicitly for each request) and replaced creation of PublicKey with Keycloak util methods already available.